dns_filter:  Kinda broken.
We have a problem where the dispatcher we use for the resolver is
created in the same process where we're doing the resolution. We don't
get an answer back until after we create the response to the client.

Changing where the DnsResolver is created gets a segfault because of an
assertion validating that the resolver is created in the same process
as the code doing the resolution.

This worked before.

Signed-off-by: Alvin Baptiste <alvinsb@gmail.com>